For larger json reads, save up to 10x memory
JSON3.jl has always been pretty loose and free w/ memory. As a default,
it just allocated a `Vector{UInt64}` with the same length as the # of
bytes in the input json string. While this heuristic works fine for
small objects, it wastes _a lot_ of memory when reading large json
files/inputs. From a random sampling of json inputs > 1000 bytes, I
found that we typically only needed a tape that was about 10% as large.

This PR implements this memory saving, only allocating a tape that is a
tenth the size of the input buffer; we then add a check to each tape
writing to ensure there is enough room. In benchmarking, this adds about
1% to total reading time, which seems a worthy cost for a 10x reduction
in memory use (for really large inputs, the memory savings probably even
makes up for the extra length checks).