Support new StructTypes.CustomStruct struct type
This allows custom unwrapping/wrapping of objects when
serializing/deserializing in a very easy way. Even the support in JSON3
is very minimal; for CustomStruct serialize, we just `lower` the custom
struct and then serialize that. When deserializing, we call `lowertype`
and deserialize whatever that type is, then pass it to
`StructTypes.construct(T, obj)`.