Treat Union{Bool, Real} as StructTypes.Struct
By default, `Union{Bool, Real}` has struct type `StructTypes.NumberType`
since `Bool <: Integer`, but when parsing a `Bool` or a `Real` from
json, we want to treat them distinctly, by parsing a `Bool` _or_ a
`Real`. The solution here is to just intercept the `read` call for those
`Union` types and pass it to the `Struct` method.